refactor: move filter bar models to core module
Move `FilterBarComponent` and `FilterBarConfiguration`
definitions from the Settings view into the core Model
folder.
